requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
ibm-cloud-sdk-core==3.18.0
ibm-watson==7.0.1
idna==3.6
install==1.3.5
paho-mqtt==1.6.1
Pillow==9.5.0
pkg_resources==0.0.0
playsound==1.3.0
pycairo==1.23.0
PyGObject==3.44.2
PyJWT==2.8.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==1.26.18
websocket-client==1.6.1
